test: replay PDF region extraction families
